Successfully posts new purchase log, purchaseId working as UUID, required UUID node module in server.js, removed createdAt and updatedAt from tables
