+++ Public/cart.js
@@ -1,23 +1,23 @@
 
-$(document).ready(function () {
-    //************* post purchase log on checkout
+// $(document).ready(function () {
+//     //************* post purchase log on checkout
 
-    $("#placeOrder").on("click", function () {
-        for (k < 0; k < localStorage.length; k++) {
-            var cartItemId = localStorageItems[i].id;
-            var newLog = {
-                purchaseId: Date.UTC(),
-                itemId: cartItemId,
-                quantity: 1,
-                date: Date.now()
-            }
-            $.post("/api/purchaseLog", newLog, function (data) {
-                console.log(data);
-                console.log("purchageLog created");
-            })
-        }
-    })
-})
+//     // $("#placeOrder").on("click", function () {
+//     //     for (var k = 0; k < localStorage.length; k++) {
+//     //         var cartItemId = localStorageItems[i].id;
+//     //         var newLog = {
+//     //             purchaseId: Date.UTC(),
+//     //             itemId: cartItemId,
+//     //             quantity: 1,
+//     //             date: Date.now()
+//     //         }
+//     //         $.post("/api/purchaseLog", newLog, function (data) {
+//     //             console.log(data);
+//     //             console.log("purchageLog created");
+//     //         })
+//     //     }
+//     // })
+// })
 
 // ************************************************************************************************************
 // ************************************************************************************************************
