+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": "Volund56",
+    "password": "",
     "database": "athenaDB",
     "host": "127.0.0.1",
     "dialect": "mysql"
