+++ package.json
@@ -26,6 +26,7 @@
     "mysql": "^2.15.0",
     "mysql2": "^1.5.3",
     "path": "^0.12.7",
-    "sequelize": "^4.37.10"
+    "sequelize": "^4.37.10",
+    "uuid": "^3.2.1"
   }
 }
